+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-## v1.2.2
-* Added the option to set the maximum amount of Vain Shrouds that can exist (defaults to false)
+## v1.3.0
+* Added a feature to bring back the Vain Shrouds in v60 (Requested by [MiraakThuri](https://github.com/EssieFir/SoberShip/issues/2))
+* Added a feature to prevent the Vain Shrouds being reset when they fail to spawn.
+* Added a feature to cap Vain Shroud iterations if it's too high.
+* Updated config descriptions to describe what they do in v60.
 
 ## v1.2.1
 * Fixed an issue where the level would never start if a non-host had the mod installed and vain shrouds tried to spawn.
@@ -20,3 +23,6 @@
 
 ## v1.1.0
 * Now removes newly generated Vain Shrouds as well.
+
+## v1.0.0
+* Created the mod